pkg/gtypes: rename Geode types package to gtypes
This is to avoid confusion with the ir/types package.
