+++ pkg/ast/DotReference.go
@@ -3,7 +3,7 @@ package ast
 import (
 	"fmt"
 
-	gtypes "github.com/geode-lang/geode/pkg/types"
+	"github.com/geode-lang/geode/pkg/gtypes"
 	"github.com/llir/llvm/ir"
 	"github.com/llir/llvm/ir/constant"
 	"github.com/llir/llvm/ir/types"
