+++ pkg/gtypes/struct.go
@@ -1,4 +1,4 @@
-package types
+package gtypes
 
 import (
 	"github.com/llir/llvm/ir/types"
